% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{add-dictionary}
\alias{add-dictionary}
\alias{copy_dict_template}
\alias{add_dictionary}
\title{Add a dictionary}
\usage{
copy_dict_template(dest_path = ".")

add_dictionary(dict_path, set_active = TRUE)
}
\arguments{
\item{dest_path}{The destination path where the dictionary template should be copied.}

\item{dict_path}{The path to the dictionary (\emph{.csv}) to add.}

\item{set_active}{if \code{TRUE}, the added dictionary becomes the active dictionary.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A dictionary is used to interprete Orphanet concepts, given as ids.
Copy the dictionary template then add it to the package after modification.

Once added, the dictionary will appear among the available options
through the \code{\link[=orphatools_options]{orphatools_options()}} interface. It can also be manually set using the built-in
\code{\link[=options]{options()}} function and the \code{"nomenclature_version"} name.
}
\seealso{
\code{\link[=orphatools_options]{orphatools_options()}}
}
