% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{add_associated_genes}
\alias{add_associated_genes}
\title{Add ORPHAcodes-genes associations to \code{orphatools}.}
\usage{
add_associated_genes(filepath, set_active = TRUE)
}
\arguments{
\item{filepath}{An xml file published by Orphanet.}

\item{set_active}{if \code{TRUE}, the added file becomes the active file.}
}
\description{
This function analyzes the file containing associations between ORPHAcodes and genes
and saves them internally in an R-friendly format.
}
\details{
Orphanet publishes the nomenclature pack on a 6-months basis, and is deployed in english only.
The different file versions are available at https://www.orphadata.com/genes/.
They must be downloaded locally to be added to \code{orphatools}.

As a gene mutation may have various effects on individuals health, this file indicates how genes and
clinical entities are related, giving the association type and status. Genes can be referred through
multiple referentials including HGNC, OMIM, Ensembl, Genatlas, Reactome and SwissProt.

The added file version will appear among the available options
through the \code{\link[=orphatools_options]{orphatools_options()}} interface. It can also be manually set using the built-in
\code{\link[=options]{options()}} function and the \code{"gene_file_version"} name.
}
\seealso{
\code{\link[=orphatools_options]{orphatools_options()}}, \code{\link[=load_associated_genes]{load_associated_genes()}}, \code{\link[=load_genes_synonyms]{load_genes_synonyms()}}
}
