% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{add_nomenclature_pack}
\alias{add_nomenclature_pack}
\title{Add a nomenclature pack to \code{orphatools}}
\usage{
add_nomenclature_pack(
  orpha_pack_nomenclature_path,
  nomenclature_xml_file,
  classifications_folder,
  set_active = TRUE,
  force = FALSE
)
}
\arguments{
\item{orpha_pack_nomenclature_path}{The path to the uncompressed nomenclature pack.}

\item{nomenclature_xml_file}{The name of the nomenclature file (\emph{.xml}).}

\item{classifications_folder}{The name of the folder containing the 34 \emph{.xml} classification files.}

\item{set_active}{if \code{TRUE}, the added pack becomes the active pack.}

\item{force}{if \code{TRUE}, adds the pack even if an identical version was internally found.}
}
\description{
This function analyzes the nomenclature pack files and saves them internally in an R-friendly format.
}
\details{
Orphanet publishes the nomenclature pack on a yearly basis, and is deployed in different language versions.
The different nomenclature pack versions are available at https://www.orphadata.com/pack-nomenclature/.
They must be downloaded locally and uncompressed to be added to \code{orphatools}.

The uncompressed Orphanet nomenclature pack contains a set of \emph{.xml} and \emph{.xlsx} files for coding, including the nomenclature file
(e.g. \emph{ORPHAnomenclature_fr_2023.xml}) and a set of more than 30 classifications (usually contained in \emph{Classifications} folder).

The nomenclature file contains for each ORPHAcode the associated properties,
like their associated label (and synonyms), status (active or not),
classification levels (group of disorder, disorder or subtype of disorder),
and redirections to another ORPHAcode (for deprecated or obsolete only).

Each classification contains numerous from/to relationships between clinical entities,
depicting the global Orphanet classification system.

Once added, the nomenclature pack will appear among the available options
through the \code{\link[=orphatools_options]{orphatools_options()}} interface. It can also be manually set using the built-in
\code{\link[=options]{options()}} function and the \code{"nomenclature_version"} name.
}
\seealso{
\code{\link[=orphatools_options]{orphatools_options()}} to switch from a pack to another,
\code{\link[=load_nomenclature]{load_nomenclature()}}, \code{\link[=load_raw_nomenclature]{load_raw_nomenclature()}},
\code{\link[=load_classifications]{load_classifications()}}, \code{\link[=load_synonyms]{load_synonyms()}}, \code{\link[=load_redirections]{load_redirections()}}
to load the data added through this function.
}
