% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{add_superNode}
\alias{add_superNode}
\title{Add an extra node above the roots of the graph}
\usage{
add_superNode(graph)
}
\arguments{
\item{graph}{The graph to add the extra node to}
}
\value{
The new graph with the added node
}
\description{
Add an extra node above the roots of the graph
}
