% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{apply_indent}
\alias{apply_indent}
\title{Shift some rows (indentation) on the base of the given index to make some hierarchical structure appear.}
\usage{
apply_indent(df, cols_to_display)
}
\arguments{
\item{df}{A sorted data.frame with ORPHA codes and their corresponding index (see \emph{assign_indent_index} function)}

\item{cols_to_display}{The columns to keep and to shift}
}
\value{
A matrix with the right indentations applied on the requested columns
}
\description{
Shift some rows (indentation) on the base of the given index to make some hierarchical structure appear.
}
\examples{
df = get_descendants('307711')
df_index = assign_indent_index(df)
df_test = df_index
df_test$N = 1:nrow(df_test)
M_shifted = apply_indent(df_test, cols_to_display = c('code', 'N'))
}
