% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{apply_orpha_indent}
\alias{apply_orpha_indent}
\title{Indentation}
\usage{
apply_orpha_indent(
  df,
  df_classif = NULL,
  indented_cols = NULL,
  prefix = "indent"
)
}
\arguments{
\item{df}{An \code{\link[=orpha_df]{orpha_df()}} instantiation.}

\item{df_classif}{The classification data to consider. If NULL, find all classifications containing the given ORPHAcode.}

\item{indented_cols}{The columns that need to be shifted.}

\item{prefix}{The prefix of the indented columns. Default is \code{"indent"}.}
}
\value{
A matrix with the right indentations applied on the requested columns
}
\description{
In order to make it easier to visualize which ORPHAcodes contain the others in a data.frame,
this function apply indentation, which means the associated ORPHAcode is one level lower in Orphanet classification than the ORPHAcode above.

ORPHAcodes may appear in multiple rows because they can be child of several parents.
}
\examples{
library(dplyr)
orpha_codes = get_descendants('307711', output='codes_only')

df = orpha_df(data.frame(orpha_code=orpha_codes), orpha_code_col='orpha_code') \%>\%
  left_join(load_nomenclature(), by='orpha_code')
df_indented = apply_orpha_indent(df, indented_cols='label')
}
