% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{assign_indent_index}
\alias{assign_indent_index}
\title{Use ORPHA classification to build hierarchical style representation of a given ORPHA tree.}
\usage{
assign_indent_index(df, current_code = NULL, current_index = "")
}
\arguments{
\item{df}{The given ORPHA tree. It is supposed to be a from/to data.frame object.}

\item{current_code}{An ORPHA code to analyse the children from. Used for recursivity.}

\item{current_index}{The index of the given ORPHA code. It will be expanded for each child. Used for recursivity.}
}
\value{
A sorted data.frame containing ORPHA codes present in the given ORPHA tree and their associated index.
This makes appear the hierarchical structure of the classification as one can see
on www.orpha.net.
}
\description{
Use ORPHA classification to build hierarchical style representation of a given ORPHA tree.
}
\examples{
df = get_descendants('307711')
df_index = assign_indent_index(df)
}
