% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{classifications}
\alias{classifications}
\alias{load_classifications}
\alias{is_in_classif}
\title{Orphanet classifications}
\usage{
load_classifications()

is_in_classif(orpha_code, df_classif)
}
\arguments{
\item{orpha_code}{An ORPHAcode.}

\item{df_classif}{A from/to data.frame providing relationships between ORPHAcodes.}
}
\value{
All classifications as a list of data.frame

Returns TRUE if a given ORPHAcode belongs to a given classification, FALSE else.
}
\description{
The functions presented below analyze relationships between ORPHAcodes and returns
results in a convenient format for further analysis or display.
}
\details{
Orphanet organized its classification system around 34 different hierarchies,
which are not completely distinct.

For each hierarchy is listed a set of from/to relationships between ORPHAcodes.
A child is always more specific than its parent.

\code{load_classifications}
Mind using \code{bind_rows} followed by \code{distinct} to merge all classification.

The classification system will change according to the chosen \code{"orphatools_nomenclature"}
option, that you can set via the \code{\link[=orphatools_options]{orphatools_options()}} interface. Add a new available option
with \code{\link[=add_nomenclature_pack]{add_nomenclature_pack()}}.
}
\examples{
library(dplyr)

all_classif = load_classifications()
df_all_classif = load_classifications() \%>\% bind_rows() \%>\% distinct()

is_in_classif(303, all_classif[[1]])
is_in_classif(303, all_classif[[6]])

}
\seealso{
\code{\link[=add_nomenclature_pack]{add_nomenclature_pack()}}, \code{\link[=orphatools_options]{orphatools_options()}}
}
