% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{classNode2code}
\alias{classNode2code}
\title{Convert an xml2 "ClassificationNode" object as a numerical Orpha code}
\usage{
classNode2code(classNode)
}
\arguments{
\item{classNode}{The xml2 "ClassificationNode" object}
}
\value{
The associated Orpha code
}
\description{
Convert an xml2 "ClassificationNode" object as a numerical Orpha code
}
