% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{code2classNode}
\alias{code2classNode}
\title{Convert a numerical Orpha code as an xml2 "ClassificationNode" object}
\usage{
code2classNode(orphaCode, class_data, all = FALSE)
}
\arguments{
\item{orphaCode}{The numerical Orpha code}

\item{class_data}{The classification data to consider}

\item{all}{If TRUE, returns all classification nodes associated to the given code}
}
\value{
An xml2 "ClassificationNode" object, as it is described in Orphanet nomenclature pack classifications
}
\description{
Convert a numerical Orpha code as an xml2 "ClassificationNode" object
}
