% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{color-graphs}
\alias{color-graphs}
\alias{color_codes}
\alias{color_class_levels}
\title{Color graphs}
\usage{
color_codes(graph, orpha_codes)

color_class_levels(graph)
}
\arguments{
\item{graph}{The graph to color.}

\item{orpha_codes}{The nodes to color.}
}
\value{
The colored graph
}
\description{
\code{color_codes} colors the given ORPHAcodes to emphasize them among the others.

\code{color_classif_levels} colors all vertices of the given graph to distinguish classification levels.
}
\examples{
# Build graph
init_codes = c(303, 305)
graph = merge_branches(init_codes, direction='down', output='graph')
plot(graph)

# Emphasize some specific ORPHAcodes
graph = color_codes(graph, init_codes)
plot(graph)

# Distinguish classification levels
graph = color_class_levels(graph)
plot(graph)
}
