% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{color_graph}
\alias{color_graph}
\title{Color vertices of a graph to emphasize some specific nodes or make classification levels appear}
\usage{
color_graph(graph, emphasize_nodes = NULL, display_class_levels = TRUE)
}
\arguments{
\item{graph, }{The graph to color}

\item{emphasize_nodes, }{List of codes to emphasize in the graph (red color).}

\item{display_class_levels}{If TRUE, color vertex frames with the associated
color (blue for groups, green for disorders and purple for subtypes)}
}
\value{

}
\description{
Color vertices of a graph to emphasize some specific nodes or make classification levels appear
}
\examples{
code = 303
codes_list = c(303, 79361, 158676, 89843)
graph = get_common_graph(code, what='both')
graph = color_graph(graph)
graph = color_graph(graph, emphasize_nodes = codes_list)
graph = color_graph(graph, emphasize_nodes = codes_list, display_class_levels = FALSE)
}
