% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{complete_family}
\alias{complete_family}
\title{This function basically finds the siblings of each provided code before building
the graph of the completed family. It is recommended to use the \code{color_graph}
function afterwards to visualize which codes were added.}
\usage{
complete_family(codes_list, class_data = NULL, include_ancestors = T)
}
\arguments{
\item{codes_list}{The codes list to extend}

\item{class_data}{The classification to consider}
}
\value{
The given codes and their relevant relatives according to the classification
}
\description{
This function basically finds the siblings of each provided code before building
the graph of the completed family. It is recommended to use the \code{color_graph}
function afterwards to visualize which codes were added.
}
\examples{
codes_list = codes_list = c('303', '305', '595356')
all_class = load_classifications()

new_graph = complete_family(codes_list)
new_graph = complete_family(codes_list, class_data = all_class[['ORPHAclassification_187_rare_skin_diseases_fr']])
}
