% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{find_roots}
\alias{find_roots}
\title{Find the root or the roots of a given graph.}
\usage{
find_roots(graph, first_only = FALSE)
}
\arguments{
\item{graph}{The graph (igraph object) to find the roots from}

\item{first_only}{If TRUE, return one root only}
}
\value{
The root(s) of the given graph as a numeric code
}
\description{
Find the root or the roots of a given graph.
}
\examples{
edgelist = get_ancestors('303')
graph = igraph::graph_from_data_frame(edgelist)
roots = find_roots(graph)

}
