% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonyms.R
\name{get_all_labels}
\alias{get_all_labels}
\title{Disease labels}
\usage{
get_all_labels(orpha_code)
}
\arguments{
\item{orpha_code}{An ORPHAcode.}
}
\value{
A character vector containing all labels associated to the given ORPHAcode.
}
\description{
Extract the synonyms of the given ORPHAcode as well as its preferential label, given in first position.
}
\seealso{
\code{\link[=load_synonyms]{load_synonyms()}} to load the synonyms table.
}
