% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{get_classification_level}
\alias{get_classification_level}
\title{Get the classification level of a given ORPHAcode}
\usage{
get_classification_level(code)
}
\arguments{
\item{orphaCode, }{ORPHAcode to get the classification level from}
}
\value{
The corresponding classification level
}
\description{
Get the classification level of a given ORPHAcode
}
