% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{get_common_graph}
\alias{get_common_graph}
\title{Get a base graph from a codes list, which is basically a graph including the relevant ancestors and the given codes.}
\usage{
get_common_graph(codes_list, class_data = NULL, colored = FALSE)
}
\arguments{
\item{codes_list}{The codes to build the base graph from}

\item{class_data}{The Orpha classification to consider to find the ancestors.
If NULL, automatically search a classification for each code}
}
\value{
The built base graph
}
\description{
Get a base graph from a codes list, which is basically a graph including the relevant ancestors and the given codes.
}
\examples{
all_class = load_classifications()
graph = get_common_graph(c(303,304,305), all_class[[6]])
}
