% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{get_descendants}
\alias{get_descendants}
\title{Get descendants recusevily from a given Orpha code}
\usage{
get_descendants(orphaCode, class_data = NULL, codes_only = FALSE, ...)
}
\arguments{
\item{orphaCode}{The numerical Orpha code}

\item{class_data}{The classification data to consider. If NULL, find one automatically. Descendance is independant from classification.}

\item{codes_only}{If FALSE, returns a data.frame object (columns "from" and "to") ready to be converted into a graph}

\item{...}{Additional parameters (edge_from_leaves, depth, return_depth) :
\itemize{
\item edge_from_leaves : The leaves (last descendants) also belong to the "from" column
\item depth : Depth information to transmit and increment recursevily
\item return_depth : Return depth for each edge in the final data.frame
}}
}
\value{
A from/to data.frame ready to be converted into a graph, or codes only if asked so
}
\description{
Get descendants recusevily from a given Orpha code
}
