% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{get_disorder_name}
\alias{get_disorder_name}
\title{Find the disorder name related to a given Orpha code}
\usage{
get_disorder_name(orphaCode, nom_data = NULL)
}
\arguments{
\item{orphaCode}{The Orpha code to consider}

\item{nom_data}{The nomenclature data as it is in the Orphanet nomenclature pack. If NULL, the function loads it itself.}
}
\value{
The disorder name related to the given orphaCode
}
\description{
Find the disorder name related to a given Orpha code
}
\examples{
orphaCode = '303'
nom_data = load_nomenclature()
disorderName = get_disorder_name(orphaCode, nom_data=nom_data)
}
