% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{get_LCAs}
\alias{get_LCAs}
\title{Extract the lowest common ancestors (LCAs) of a codes list in a given graph which was previously calculated.
Discussion : When there are two independent LCAs, we don't compare how "high" there are compared to the given nodes.}
\usage{
get_LCAs(codes_list, class_data = NULL)
}
\arguments{
\item{codes_list}{The codes list to extract the LCAs from}

\item{class_data}{The Orpha classification to consider to find the ancestors.
If NULL, automatically search a classification for each code}
}
\value{
The lowest common ancestors of the given codes
}
\description{
Extract the lowest common ancestors (LCAs) of a codes list in a given graph which was previously calculated.
Discussion : When there are two independent LCAs, we don't compare how "high" there are compared to the given nodes.
}
\examples{
codes_list = c(303, 305, 595356)

get_LCAs(codes_list)
get_LCAs(codes_list, class_data = all_class[['ORPHAclassification_187_rare_skin_diseases_fr']])
}
