% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{get_lowest_groups}
\alias{get_lowest_groups}
\title{Analyze ancestors to find the lowest group containing the given code}
\usage{
get_lowest_groups(orphaCode, class_data = NULL)
}
\arguments{
\item{orphaCode}{The code to find the lowest group from}

\item{class_data}{The classification data to consider.
If NULL, find all classifications containing the given Orpha code.
Ancestors may differ depending on the chosen classification.}
}
\value{
The lowest group containing the code, or the code itself if it is a group of disorders already
}
\description{
Analyze ancestors to find the lowest group containing the given code
}
\examples{
codes_list = c(303, 305, 595356)
get_lowest_groups(orphaCode = 158676)
}
