% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{get_relatives}
\alias{get_relatives}
\title{Find relationships between the given codes and tries to extend it with relevant ones.}
\usage{
get_relatives(
  codes_list,
  class_data = NULL,
  colored = TRUE,
  children_only = FALSE
)
}
\arguments{
\item{codes_list}{The codes list to extend}

\item{class_data}{The classification to consider}

\item{colored}{If TRUE, nodes are colored with green for the given codes and with red for the codes found as relatives}

\item{children_only}{If TRUE, no ancestors are searched and the given codes are considered as the graph roots.
If FALSE, roots are the LCAs of the given codes. When the classification head is found as LCA, user may experiment calculation issues.}

\item{base_graph}{The graph to base the calculation on}
}
\value{
The given codes and their relevant relatives according to the classification
}
\description{
Find relationships between the given codes and tries to extend it with relevant ones.
}
