% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{graph-operations}
\alias{graph-operations}
\alias{find_roots}
\alias{find_leaves}
\title{Operations on graphs}
\usage{
find_roots(graph)

find_leaves(graph)
}
\arguments{
\item{graph}{An igraph object.}
}
\value{
The extrem nodes of a graph.
}
\description{
Orphanet classification system can be depicted by oriented graphs as it is a wide list of
parent/child relationships.

The \code{igraph} package provides useful tools for graphs analysis that we can use for the
study of rare diseases.

The roots are the ORPHAcodes without any parent,
like the heads of classification. The leaves are the ORPHAcodes without any children.
They are usually disorders or subtypes of disorders (see \code{\link[=get_classification_level]{get_classification_level()}}).
}
\examples{
graph = get_ancestors('303', output='graph')

roots = find_roots(graph)
leaves = find_leaves(graph)

}
