% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplication.R
\name{group_by_code}
\alias{group_by_code}
\title{A common group_by operation to take descendants of the present Orpha codes into account in the aggregation.
Aggregation works if column to group is named "code".}
\usage{
group_by_code(.data, ..., class_data = NULL, include_descendants = TRUE)
}
\arguments{
\item{.data}{The data consider}

\item{...}{Additional parameters to transmit to the dplyr group_by function}

\item{class_data}{The classification data to consider. If NULL, automatically search a classification for each code}

\item{include_descendants}{}
}
\value{
Results of the group_by operation
}
\description{
A common group_by operation to take descendants of the present Orpha codes into account in the aggregation.
Aggregation works if column to group is named "code".
}
