% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{in_between_graph}
\alias{in_between_graph}
\title{Extract a graph from a set of ORPHAcodes}
\usage{
in_between_graph(vs, df_classif = NULL)
}
\arguments{
\item{vs}{A set of ORPHAcodes used to extract the graph.}

\item{df_classif}{The classification to consider. If NULL, loads the whole Orphanet classification.}
}
\value{
The extracted graph as an \emph{igraph} object.
}
\description{
The extracted graph contains all the ORPHAcodes specified in the \code{vs} argument.
It may include other ORPHAcodes inside, but all the roots and leaves are part of \code{vs}.
}
