% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{is_in_classif}
\alias{is_in_classif}
\title{Checks if a given Orpha code belongs to a given classification.}
\usage{
is_in_classif(orphaCode, class_data)
}
\arguments{
\item{orphaCode}{The given Orpha code}

\item{class_data}{One of the 33 Orpha classifications}
}
\value{

}
\description{
Checks if a given Orpha code belongs to a given classification.
}
\examples{
all_class = load_classifications()
print(is_in_classif('303', all_class[[1]]))
print(is_in_classif('303', all_class[[6]]))
}
