% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_custom.R
\name{layout_custom}
\alias{layout_custom}
\title{Calculate nodes position for a fancy layout}
\usage{
layout_custom(graph, reverse_y = TRUE, factor = 5)
}
\arguments{
\item{graph}{The graph to display}

\item{reverse_y}{If TRUE, display from top to bottom}

\item{factor}{(default 5) This parameter tells how far the nodes should be from each other}
}
\value{
An adapted layout to plot nodes from Orpha classifications
}
\description{
Calculate nodes position for a fancy layout
}
