% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{load-genes}
\alias{load-genes}
\alias{load_associated_genes}
\alias{load_genes_synonyms}
\title{Associated genes}
\usage{
load_associated_genes()

load_genes_synonyms()
}
\value{
A data frame containing for each ORPHAcode the kind of association (if any).
with a related gene and all known information (symbol, name, references, locus) about the latter.
}
\description{
Load all needed information to explore associations between ORPHAcodes and genes.

\code{load_associated_genes} loads the association tables.

\code{load_gene_synonyms} loads genes synonyms in case your available data refers
to genes in a different way.
}
\details{
The loaded dataframes directly depend on the \code{"orphatools_gene_file"} option that
you can set via the \code{\link[=orphatools_options]{orphatools_options()}} interface. You can add an available option
with \code{\link[=add_associated_genes]{add_associated_genes()}}.
}
\examples{
df_associated_genes = load_associated_genes()
df_genes_synonyms = load_genes_synonyms()

}
\seealso{
\code{\link[=orphatools_options]{orphatools_options()}}, \code{\link[=add_associated_genes]{add_associated_genes()}}, \code{\link[=specify_code]{specify_code()}}
}
