% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{load_class_levels}
\alias{load_class_levels}
\title{Generate and load an OrphaCode -> Classification Level table}
\usage{
load_class_levels(load = TRUE, save = TRUE)
}
\arguments{
\item{load}{Load a dumped output. It the output does not exist, it is recalculated}

\item{save}{Dump the output for later readings}
}
\value{
The OrphaCode -> Classification Level table, as defined in the Orphanet nomenclature
}
\description{
Generate and load an OrphaCode -> Classification Level table
}
\examples{
class_levels = load_class_levels()

}
