% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{load_classifications}
\alias{load_classifications}
\title{Load the 33 classifications given in the Orphanet nomenclature pack}
\usage{
load_classifications(load = TRUE, save = TRUE)
}
\arguments{
\item{load}{Load a dumped output. It the output does not exist, it is recalculated.}

\item{save}{Dump the output for later readings}
}
\value{
All classifications as xml2 objects
}
\description{
Load the 33 classifications given in the Orphanet nomenclature pack
}
\examples{
all_class = load_classifications()
}
