% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{load_genes_synonyms}
\alias{load_genes_synonyms}
\title{Load all possible synonyms for each known gene in Orphanet public data}
\usage{
load_genes_synonyms()
}
\value{
A data frame containing genes ids for joining operations, their full name and main symbol and all possible synonyms
}
\description{
Load all possible synonyms for each known gene in Orphanet public data
}
