% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{load_nomenclature}
\alias{load_nomenclature}
\title{Load the nomenclature file given in the Orphanet nomenclature pack}
\usage{
load_nomenclature()
}
\value{
The nomenclature data as an xml2 object
}
\description{
Load the nomenclature file given in the Orphanet nomenclature pack
}
