% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{load_orpha_genes}
\alias{load_orpha_genes}
\title{Load the list of genes provided by Orphanet and their associated ORPHAcodes}
\usage{
load_orpha_genes()
}
\value{
A data frame containing for each ORPHAcode the kind of association (if any)
with a related gene and all known information (symbol, name, references, locus) about the latter.
}
\description{
Load the list of genes provided by Orphanet and their associated ORPHAcodes
}
