% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirections.R
\name{load_redirections}
\alias{load_redirections}
\alias{load_raw_redirections}
\title{Redirections}
\usage{
load_redirections()

load_raw_redirections()
}
\value{
A data.frame object giving the deprecated or obsolete ORPHAcodes,
the corresponding associations with the association type
}
\description{
Load for each obsolete or deprecated ORPHAcode the corresponding association (\emph{Moved to} or \emph{Referred to}).

While \code{load_raw_redirections} keep the original id values for each Orphanet concept,
\code{load_redirections} translate them according to \code{"orphatools_dict"} option, which can be set manually using
built-in \code{\link[=options]{options()}} function or through the \code{\link[=orphatools_options]{orphatools_options()}} interface.
}
\examples{
df_redirections = load_raw_redirections()
df_redirections = load_redirections()
}
\seealso{
\code{\link[=redirect_code]{redirect_code()}}
}
