% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonyms.R
\name{load_synonyms}
\alias{load_synonyms}
\title{Synonyms}
\usage{
load_synonyms()
}
\value{
A 3-columns data.frame indicating for each ORPHAcode its preferential label and synonyms.
\code{NA} is set if no synonyms was found.
}
\description{
Any disease can be named in multiple ways. For each ORPHAcode, Orphanet provides a
unique preferential label. Any other name is then called synonym.

This function returns in a data.frame the ORPHAcodes, their preferential label and synonyms if any.
}
\seealso{
\code{\link[=get_all_labels]{get_all_labels()}} to directly extract all the names associated to the given ORPHAcode.
}
