% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplication.R
\name{mutate}
\alias{mutate}
\title{Overridden version of \code{dplyr::mutate} for dataframes grouped
with \code{group_by_code} function.}
\usage{
mutate(df, ...)
}
\arguments{
\item{df}{Dataframe to mutate}

\item{...}{arguments for dplyr::mutate}
}
\value{
The mutated dataframe
}
\description{
Overridden version of \code{dplyr::mutate} for dataframes grouped
with \code{group_by_code} function.
}
\examples{
library(magrittr)
df = data.frame(code = c('303', '79408', '79408'), val = 1:3)
# df \%>\% group_by_code \%>\% dplyr::mutate(val2 = max(val)) # May rise an error "Can't recycle..."
df \%>\% group_by_code \%>\% orphatools::mutate(val2 = max(val)) # Works well
}
