% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplication.R
\name{orpha_df}
\alias{orpha_df}
\title{The orpha_df class allows taking ORPHA structure when codes are grouped together.
For instance, the counting of individuals related to a specific disorder will also
take into account individuals having the related subtypes of this disorder.
It needs the specification of the column of x containing the ORPHAcodes}
\usage{
orpha_df(x, orphaCode_col = "diag_code", class_data = NULL, force_nodes = NULL)
}
\arguments{
\item{x}{A data frame or data frame extension}

\item{orphaCode_col}{The column name of \code{x} containing the ORPHAcodes}

\item{force_nodes}{}
}
\value{
An orpha_df instantiation
}
\description{
The orpha_df class allows taking ORPHA structure when codes are grouped together.
For instance, the counting of individuals related to a specific disorder will also
take into account individuals having the related subtypes of this disorder.
It needs the specification of the column of x containing the ORPHAcodes
}
