% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{orphacode-properties}
\alias{orphacode-properties}
\alias{get_label}
\alias{get_classification_level}
\alias{get_status}
\alias{is_active}
\alias{get_type}
\title{ORPHAcode specific properties}
\usage{
get_label(orpha_codes)

get_classification_level(orpha_codes)

get_status(orpha_codes)

is_active(orpha_codes)

get_type(orpha_codes)
}
\arguments{
\item{orpha_codes}{The ORPHAcodes as a vector from which properties should be returned.}
}
\value{
The corresponding properties of the given ORPHAcodes.
}
\description{
Get specific properties of given ORPHAcodes.
}
\details{
The label is the clinical entity name and is returned by \code{get_label}.
As multiple synonyms can be associated to a clinical entity, this function only returns the preferrential labels.
To get all synonyms instead, use \code{get_synonyms}. Besides, if the displayed language does not meet your needs,
you should consider changing options via \code{\link[=orphatools_options]{orphatools_options()}} or adding a new nomenclature pack using \code{\link[=add_nomenclature_pack]{add_nomenclature_pack()}}.

The classification level is the clinical entity name and is returned by \code{get_classification_level}.
Classification level can either be a group of disorders, a disorder or a subtype of disorder.

The status of an ORPHAcode is a value contained in the following terms :
"Active", "Inactive: Deprecated", "Inactive: Obsolete", "Inactive: Non rare disease in Europe".
It is returned by \code{get_status}, while \code{is_active} just returns TRUE if status is "Active" and FALSE else.

The disorder type indicates the clinical entity's typology and is returned by \code{get_type}.

For more details, please read the full Orphanet description included in the nomenclature pack (\emph{.pdf}).
}
\note{
All properties described above are accessible in a combined format through the \code{load_nomenclature} function.
}
\examples{
orpha_codes = c('303', '595356', '79410')
get_label(orpha_codes)
get_classification_level(orpha_codes)
get_status(orpha_codes)
is_active(orpha_codes)
get_type(orpha_codes)

}
\seealso{
\code{\link[=load_nomenclature]{load_nomenclature()}}, \code{\link[=load_raw_nomenclature]{load_raw_nomenclature()}}
}
