% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orphatools-package.R
\docType{package}
\name{orphatools-package}
\alias{orphatools}
\alias{orphatools-package}
\title{orphatools: Tools to deal with the Orpha classifications}
\description{
Use the orpha classifications to find ancestors, children or even relatives of an orpha code.
}
\author{
\strong{Maintainer}: Remy Dumas \email{remy.dumas@aphp.fr} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
