% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{orphatools_options}
\alias{orphatools_options}
\title{Options for \code{orphatools}}
\usage{
orphatools_options()
}
\description{
Interface to list and set orphatools-related options.

Options are useful to roll back to older versions or to switch between languages for display.
\itemize{
\item \code{"orphatools_dict"} refers to the dictionary used to interprete Orphanet concepts ids.
\item \code{"orphatools_nomenclature"} refers to a nomenclature pack version (extraction date and used language).
\item \code{"orphatools_gene_file"} refers to the version of the file used to associated genes to ORPHAcodes.
}

All options can be set manually via the built-in \code{\link[=options]{options()}} function.
}
\examples{
\dontrun{orphatools_options()}
}
\seealso{
\code{\link[=add_dictionary]{add_dictionary()}}, \code{\link[=add_nomenclature_pack]{add_nomenclature_pack()}}, \code{\link[=add_associated_genes]{add_associated_genes()}}
}
