% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirections.R
\name{redirect_code}
\alias{redirect_code}
\title{Apply ORPHAcodes redirections.}
\usage{
redirect_code(orpha_codes, deprecated_only = TRUE)
}
\arguments{
\item{orpha_codes}{ORPHAcodes to redirect.}

\item{deprecated_only}{If \code{TRUE}, redirect only deprecated ORPHAcodes.}
}
\value{
The redirected ORPHAcodes. If no redirection is found, ORPHAcodes remain the same as given,
so ORPHAcodes may remain inactive after redirection.
}
\description{
An ORPHAcode is redirected to an active ORPHAcode when it becomes deprecated or obsolete,
respectively with a \emph{moved to} (appropriate redirection) or \emph{referred to} (suggestion) association type.

There is no redirection when status is \emph{"Inactive: Non rare disease in Europe"}.
Redirection is always provided for deprecated ORPHAcodes, but not for obsolete ones.
}
\examples{
orpha_codes = c(303, 166068, 166457)
redirect_code(orpha_codes)
redirect_code(orpha_codes, deprecated_only=FALSE)
}
\seealso{
\code{\link[=load_redirections]{load_redirections()}}, \code{\link[=load_raw_redirections]{load_raw_redirections()}}
}
