% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{specify_code}
\alias{specify_code}
\title{Find a more precise ORPHAcode according to the provided genes}
\usage{
specify_code(orpha_codes, hgnc_codes, by = NULL)
}
\arguments{
\item{orpha_codes}{The ORPHAcodes to update}

\item{hgnc_codes}{The provided genes}

\item{by}{This argument specifies the groups in which the provided genes should be considered.}
}
\value{
The updated ORPHAcodes, the same length as the \code{orpha_codes}.
An ORPHAcode is changed if a single correspondence was found.
}
\description{
Find a more precise ORPHAcode according to the provided genes
}
\examples{
library(dplyr)
df = tibble(
  patient_id=c('A', 'A', 'B', 'C', 'D', 'D'),
  initial_orphaCode = c("65753", "65753", "903", "65753", "65753", "65753"), # CMT1 and von Willebrand
  hgnc_code = c("16463", "16361", "15714", "26792", "16463", "15714")) # MPZ, VWF and LITAF

df \%>\% mutate(assigned_orphaCode = specify_code(initial_orphaCode, hgnc_code))
df \%>\% mutate(assigned_orphaCode = specify_code(initial_orphaCode, hgnc_code, by='patient_id'))
}
