% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{translate_class_level}
\alias{translate_class_level}
\title{Translate a classLevel column into human readable data}
\usage{
translate_class_level(df, col = "classLevel")
}
\arguments{
\item{df, }{The dataframe containing the classLevel column}

\item{col, }{To change the classLevel column name if necessary}
}
\value{
The dataframe with the translated column
}
\description{
Translate a classLevel column into human readable data
}
