% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{translate_disorder_types}
\alias{translate_disorder_types}
\title{Translate a disorderType column into human readable data}
\usage{
translate_disorder_types(V)
}
\arguments{
\item{V, }{Vector to translate}
}
\value{
The dataframe with the translated column
}
\description{
Translate a disorderType column into human readable data
}
