% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{translate_properties}
\alias{translate_properties}
\title{Translate abstract codes into literal values to make them more understandable}
\usage{
translate_properties(nom_data)
}
\arguments{
\item{nom_data}{The nomenclature data as loaded with the \code{load_nomenclature} function.}
}
\value{
The same properties (name, flag value, disorder type, classification level)
in a human-readable way
}
\description{
Translate abstract codes into literal values to make them more understandable
}
\examples{
nom_data = load_nomenclature()
nom_data_new = translate_properties(nom_data)

}
