% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{update_genes}
\alias{update_genes}
\title{Update genes information from orphanet public data,
available here: https://www.orphadata.com/genes/}
\usage{
update_genes(genes_file)
}
\arguments{
\item{genes_file}{An xml file published by Orphanet}
}
\description{
Update genes information from orphanet public data,
available here: https://www.orphadata.com/genes/
}
