% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{update_nomenclature_pack}
\alias{update_nomenclature_pack}
\title{Update the Orpha nomenclature pack. Orphanet updates its pack on a yearly basis.}
\usage{
update_nomenclature_pack(
  orpha_pack_nomenclature_path,
  nomenclature_xml_file = "ORPHAnomenclature_fr.xml",
  classifications_folder = "Classifications"
)
}
\arguments{
\item{orpha_pack_nomenclature_path}{the path to the uncompressed nomenclature pack.
You'll find it available here : https://www.orphadata.com/pack-nomenclature/}

\item{nomenclature_xml_file}{nomenclature file is supposed to be \emph{ORPHAnomenclature_fr.xml}, but might differ.}

\item{classifications_folder}{classifications folder is supposed to be \emph{Classifications_fr}, but might differ}
}
\value{

}
\description{
Update the Orpha nomenclature pack. Orphanet updates its pack on a yearly basis.
}
