% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{upper-classification-levels}
\alias{upper-classification-levels}
\alias{subtype_to_disorder}
\alias{get_lowest_groups}
\title{Find specific ancestors}
\usage{
subtype_to_disorder(orpha_code, df_classif = NULL)

get_lowest_groups(orpha_code, df_classif = NULL)
}
\arguments{
\item{orpha_code}{A vector of ORPHAcodes.}

\item{df_classif}{The classification to consider. If NULL, loads the whole Orphanet classification.}
}
\value{
\code{subtype_to_disorder} returns the associated disorder, or the ORPHAcode itself if it
is already a disorder, or NULL if it is a group of disorders.
If a vector of ORPHAcodes is provided, function is applied on each element,
and the associated vector is returned.

\code{get_lowest_groups} returns the closest groups from the given ORPHAcode,
or the ORPHAcode itself if it is a group of disorders already.
}
\description{
These functions helps you to find key ORPHAcodes which are located above the given ORPHAcode
in the classification system.
}
\examples{
subtype_to_disorder(orpha_code = '158676')
# ORPHA:158676 is a subtype of disorder

subtype_to_disorder(orpha_code = '303')
# ORPHA:303 is a group of disorder

get_lowest_groups(orpha_code = '158676')

}
