% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vertical_positions}
\alias{vertical_positions}
\title{Calculate vertical coordinates for each node of the given graph}
\usage{
vertical_positions(graph)
}
\arguments{
\item{graph}{The graph from which nodes positions (y axis) will be calculated}
}
\value{

}
\description{
Calculate vertical coordinates for each node of the given graph
}
\examples{
df_edges = data.frame(from=c('A', 'B', 'C'), to = c('B', 'C', 'D'), depth = c(-3,-2,-1))
g = igraph::graph_from_data_frame(df_edges)
df_y = vertical_positions(g)
}
