% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dict.R
\name{add-dictionary}
\alias{add-dictionary}
\alias{copy_dict_template}
\alias{add_dictionary}
\title{Add a dictionary}
\usage{
copy_dict_template(dest_path = ".")

add_dictionary(
  filepath,
  default = FALSE,
  destdir = tools::R_user_dir("orphatools", "data")
)
}
\arguments{
\item{dest_path}{The destination path where the dictionary template should be copied.}

\item{filepath}{The location of the \emph{.csv} file containing the labels in another language.
A template of such a dicitonary can be saved using \code{copy_dict_template}.}

\item{default}{If \code{TRUE}, set the added dictionary as default.}

\item{destdir}{The destination directory, in which the processed data will be saved.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A dictionary is used to interprete Orphanet concepts, given as ids.
Copy the dictionary template then add it to the package after modification.

Once added, the dictionary will appear among the available options
through the \code{\link[=orphatools_options]{orphatools_options()}} interface. It can also be manually set using the built-in
\code{\link[=options]{options()}} function and the \code{"nomenclature_version"} name.
}
\seealso{
\code{\link[=orphatools_options]{orphatools_options()}}
}
