% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pack.R
\name{add_nomenclature_pack}
\alias{add_nomenclature_pack}
\title{Add a nomenclature pack to \code{orphatools}}
\usage{
add_nomenclature_pack(
  zip_filepath,
  default = FALSE,
  force = FALSE,
  destdir = tools::R_user_dir("orphatools", "data")
)
}
\arguments{
\item{zip_filepath}{The location of the \emph{.zip} file containing the nomenclature pack.}

\item{default}{If \code{TRUE}, set the added pack as default.}

\item{force}{If \code{TRUE}, adds the pack even if an identical version was internally found.}

\item{destdir}{The destination directory, in which the processed data will be saved.}
}
\description{
This function analyzes the nomenclature pack files and saves them internally in an R-friendly format.
}
\details{
Orphanet publishes the nomenclature pack on a yearly basis, and is deployed in different language versions.
The different nomenclature pack versions are available \href{https://www.orphadata.com/pack-nomenclature/}{here}.
They must be downloaded locally and uncompressed to be added to \code{orphatools}.

The uncompressed Orphanet nomenclature pack contains a set of \emph{.xml} and \emph{.xlsx} files for coding, including the nomenclature file
(e.g. \emph{ORPHAnomenclature_fr_2023.xml}) and a set of more than 30 classifications (usually contained in \emph{Classifications} folder).

The nomenclature file contains for each ORPHAcode the associated properties,
like their associated label (and synonyms), status (active or not),
classification levels (group of disorder, disorder or subtype of disorder),
and redirections to another ORPHAcode (for deprecated or obsolete only).

Each classification contains numerous from/to relationships between clinical entities,
depicting the global Orphanet classification system.

Once added, the nomenclature pack will appear among the available options
through the \code{\link[=orphatools_options]{orphatools_options()}} interface. It can also be manually set using the built-in
\code{\link[=options]{options()}} function and the \code{"nomenclature_version"} name.
}
\seealso{
\code{\link[=orphatools_options]{orphatools_options()}} to switch from a pack to another,
\code{\link[=load_nomenclature]{load_nomenclature()}}, \code{\link[=load_raw_nomenclature]{load_raw_nomenclature()}},
\code{\link[=load_classifications]{load_classifications()}}, \code{\link[=load_synonyms]{load_synonyms()}}, \code{\link[=load_redirections]{load_redirections()}}
to load the data added through this function.
}
