% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{analyze-genealogy}
\alias{analyze-genealogy}
\alias{get_parents}
\alias{get_children}
\alias{get_ancestors}
\alias{get_descendants}
\alias{get_siblings}
\alias{merge_branches}
\alias{complete_family}
\alias{get_LCAs}
\title{Analyze ORPHA classifications}
\usage{
get_parents(orpha_code, output = "codes_only", df_classif = NULL)

get_children(orpha_code, output = "codes_only", df_classif = NULL)

get_ancestors(
  orpha_code,
  output = "codes_only",
  max_depth = NULL,
  df_classif = NULL
)

get_descendants(
  orpha_code,
  output = "codes_only",
  max_depth = NULL,
  df_classif = NULL
)

get_siblings(orpha_code, output = "codes_only", df_classif = NULL)

merge_branches(
  orpha_codes,
  output = "codes_only",
  df_classif = NULL,
  direction = c("up", "down", "both")
)

complete_family(
  orpha_codes,
  output = "codes_only",
  df_classif = NULL,
  max_depth = 1
)

get_LCAs(orpha_codes, df_classif = NULL)
}
\arguments{
\item{orpha_code}{The ORPHAcode to start from.}

\item{output}{A value specifying the output format. Should be in \code{c("codes_only", "edgelist", "graph")}.}

\item{df_classif}{The classification data to consider. If NULL, load the whole Orphanet classification.}

\item{max_depth}{The maximum reached depth starting from the given ORPHAcode.}

\item{orpha_codes}{A vector of ORPHAcodes used to perform operations.}

\item{direction}{This gives the ancestors and/or descendants direction to complete the family. Should be in \code{c("up", "down", "both")}.}
}
\value{
Results are returned in the format specified in \code{output} argument :
\code{'codes_only'}, \code{'edgelist'}, \code{'graph'}.
}
\description{
The Orphanet classification system can be seen as a wide genealogy tree, that's why
the following terms will refer to this field.

There two classes of functions:
\itemize{
\item Functions applied to a single ORPHAcode:
\itemize{
\item \code{get_parents}
\item \code{get_children}
\item \code{get_ancestors}
\item \code{get_descendants}
\item \code{get_siblings}
}
\item Functions applied to a set of ORPHAcodes:
\itemize{
\item \code{merge_branches}
\item \code{complete_family}
\item \code{get_LCAs}
}
}

When you don't want to analyze to whole Orphanet classification system, \code{df_classif} should be set.

When only ORPHAcodes are needed, without the edges information, set \code{output='codes_only'}.
Alternatively, if you wish to analyze or visualize ORPHAcodes interactions,
set \code{output='edgelist'} or \code{output='graph'}.
}
\examples{
library(dplyr)
all_classif = load_classifications()
orpha_codes = c('303', '305', '595356')
orpha_code = '303'

# ---
# Ancestors

# Get ancestors ORPHAcodes only
ancestors_codes = get_ancestors(orpha_code)

# Change output format
df_ancestors = get_ancestors(orpha_code, output='edgelist')
graph_ancestors = get_ancestors(orpha_code, output='graph')

# Get parents only
parents = get_ancestors(orpha_code, max_depth=1)
parents = get_parents(orpha_code)

# Select a specific classification tree
classif1 = all_classif[['ORPHAclassification_156_rare_genetic_diseases_fr']]
ancestors = get_ancestors(orpha_code, df_classif=classif1)

classif2 = all_classif[['ORPHAclassification_146_rare_cardiac_diseases_fr']]
ancestors = get_ancestors(orpha_code, df_classif=classif2)

# ---
# Descendants

# Get desendants ORPHAcodes only
descendants_codes = get_descendants(orpha_code)

# Change output format
df_descendants = get_descendants(orpha_code, output='edgelist')
graph_descendants = get_descendants(orpha_code, output='graph')

# Get children only
df_children = get_descendants(orpha_code, max_depth=1)
df_childen = get_children(orpha_code)

# ---
# Siblings

siblings = get_siblings(orpha_code)
df_siblings = get_siblings(orpha_code, output='edgelist')
graph_siblings = get_siblings(orpha_code, output='graph')

# ---
# Lowest common ancestors

lcas = get_LCAs(orpha_codes)

classif = all_classif[['ORPHAclassification_187_rare_skin_diseases_fr']]
lcas = get_LCAs(orpha_codes, df_classif = classif)

# ---
# Merge branches

family_codes = merge_branches(orpha_codes)
family_codes = merge_branches(orpha_codes, df_classif = all_classif[[1]])
family_codes = merge_branches(orpha_codes, direction='up')
family_codes = merge_branches(orpha_codes, direction='down')
df_family = merge_branches(orpha_codes, output='edgelist')
graph_family = merge_branches(orpha_codes, output='graph')

# ---
# Complete family
family_codes = complete_family(orpha_codes)
family_codes = complete_family(orpha_codes, df_classif=all_classif[[1]])
family_codes = complete_family(orpha_codes, max_depth=0)
df_family = complete_family(orpha_codes, output = 'edgelist')
graph_family = complete_family(orpha_codes, output='graph')

}
