% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{find_leaves}
\alias{find_leaves}
\title{Find the leaves of a given graph.}
\usage{
find_leaves(graph)
}
\arguments{
\item{graph}{The graph (igraph object) to find the leaves from}
}
\value{
The leaves of the given graph as numeric codes
}
\description{
Find the leaves of a given graph.
}
\examples{
edgelist = get_descendants('303')
graph = igraph::graph_from_data_frame(edgelist)
leaves = find_leaves(graph)

}
