% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{find_leaves}
\alias{find_leaves}
\title{Find the leaves of a given graph.}
\usage{
find_leaves(graph)
}
\arguments{
\item{graph}{The graph to find the leaves from}
}
\value{
The leaves of the given graph as numeric codes
}
\description{
Find the leaves of a given graph.
}
\examples{
all_class = load_classifications()
graph = get_common_graph(c(303,304,305), all_class[[6]])
leaves = find_leaves(graph)
}
