% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{find_roots}
\alias{find_roots}
\title{Find the root or the roots of a given graph.}
\usage{
find_roots(graph, first_only = FALSE)
}
\arguments{
\item{graph}{The graph to find the root from}

\item{first_only}{If TRUE, return one root only}
}
\value{
The root(s) of the given graph as a numeric code
}
\description{
Find the root or the roots of a given graph.
}
\examples{
all_class = load_classifications()
graph = get_common_graph(c(303,304,305), all_class[[6]])
root = find_roots(graph)
}
