% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{get_ancestors}
\alias{get_ancestors}
\title{Find all ancestors of a given Orpha code as it is in the Orphanet classification data}
\usage{
get_ancestors(
  orphaCode,
  class_data = NULL,
  codes_only = FALSE,
  full_connections = FALSE
)
}
\arguments{
\item{orphaCode}{The numerical Orpha code}

\item{class_data}{The classification data to consider. If NULL, find all classifications containing the given Orpha code. Ancestors may differ depending on the chosen classification.}

\item{codes_only}{If FALSE, returns a data.frame object (columns "from" and "to") ready to be converted into a graph}
}
\value{
A from/to data.frame ready to be converted into a graph, or codes only if asked so
}
\description{
Find all ancestors of a given Orpha code as it is in the Orphanet classification data
}
