% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{get_code_properties}
\alias{get_code_properties}
\title{Find properties related to a given Orpha code}
\usage{
get_code_properties(orphaCode, nom_data = NULL)
}
\arguments{
\item{orphaCode}{The Orpha code to consider}

\item{nom_data}{The nomenclature data as it is in the Orphanet nomenclature pack. If NULL, the function loads it itself.}
}
\value{
The properties (flag value, disorder type, classification level) associated to the given Orpha code
}
\description{
Find properties related to a given Orpha code
}
