% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{get_siblings}
\alias{get_siblings}
\title{Get siblings of an Orpha code.}
\usage{
get_siblings(orphaCode, class_data = NULL, codes_only = T)
}
\arguments{
\item{orphaCode}{The numerical Orpha code}

\item{class_data}{The classification data to consider. If NULL, find one automatically. Descendance is independant from classification.}

\item{codes_only}{If FALSE, returns a data.frame object (columns "from" and "to") ready to be converted into a graph}
}
\value{
A from/to data.frame ready to be converted into a graph, or codes only if asked so
}
\description{
Get siblings of an Orpha code.
}
\examples{
code = '303'
siblings = get_siblings(code)
}
