% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{interactive_plot}
\alias{interactive_plot}
\title{This function was designed for a better visualization of large graphs.}
\usage{
interactive_plot(graph)
}
\arguments{
\item{graph}{An igraph object to visualize}
}
\description{
This function was designed for a better visualization of large graphs.
}
\examples{
graph = get_descendants(68346)
interactive_plot(graph)
}
