% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{interactive_plot}
\alias{interactive_plot}
\title{This function was designed for a better visualization of large graphs.}
\usage{
interactive_plot(graph, layout_tree = FALSE)
}
\arguments{
\item{graph}{An igraph object to visualize}

\item{layout_tree}{if TRUE, use the layout function above to set initial nodes position}
}
\description{
This function was designed for a better visualization of large graphs.
}
\examples{
D = get_descendants('68346')
graph = igraph::graph_from_data_frame(D)
interactive_plot(graph)
}
