% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{interactive_plot}
\alias{interactive_plot}
\title{Dynamic plot}
\usage{
interactive_plot(graph, layout_tree = FALSE)
}
\arguments{
\item{graph}{The graph to visualize.}

\item{layout_tree}{If \code{TRUE}, use \code{\link[=layout_tree]{layout_tree()}} to set initial nodes position.}
}
\description{
This function uses \code{visNetwork} library for a better visualization of larger graphs.
}
\examples{
\dontrun{
graph = get_descendants('68346', output='graph')
interactive_plot(graph)
}
}
