% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{layout_tree}
\alias{layout_tree}
\title{Calculate nodes position for a fancy layout}
\usage{
layout_tree(graph, reverse_y = TRUE, factor = 5)
}
\arguments{
\item{graph}{The graph to display}

\item{reverse_y}{If TRUE, display from top to bottom}

\item{factor}{(default 5) This parameter tells how far the nodes should be from each other}
}
\value{
An adapted layout to plot nodes from Orpha classifications
}
\description{
Calculate nodes position for a fancy layout
}
\examples{
df_ancestors = get_ancestors(303)
graph = igraph::graph_from_data_frame(df_ancestors)
plot(graph)
plot(graph, layout=igraph::layout_as_tree)
plot(graph, layout=layout_tree)
}
