% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{load_nomenclature}
\alias{load_nomenclature}
\alias{load_raw_nomenclature}
\title{ORPHAcodes properties}
\usage{
load_raw_nomenclature()

load_nomenclature()
}
\value{
The ORPHAcodes properties as a data.frame object.
}
\description{
Load all ORPHAcodes properties in a combined data.frame.
}
\details{
The label is the clinical entity name and is returned in the \code{label} column.
If the displayed language does not meet your needs,
you should consider changing options via \code{\link[=orphatools_options]{orphatools_options()}} or adding a new nomenclature pack using \code{\link[=add_nomenclature_pack]{add_nomenclature_pack()}}.

The classification level is the clinical entity name and is returned in the \code{level} column.
Classification level can either be a group of disorders, a disorder or a subtype of disorder.

The status of an ORPHAcode is a value contained in the following terms :
"Active", "Inactive: Deprecated", "Inactive: Obsolete", "Inactive: Non rare disease in Europe".

The disorder type indicates the clinical entity's typology and is returned by \code{get_type}.

For more details, please read the full Orphanet description included in the nomenclature pack (\emph{.pdf}).
}
\seealso{
\code{\link[=get_label]{get_label()}}, \code{\link[=get_classification_level]{get_classification_level()}}, \code{\link[=get_status]{get_status()}}, \code{\link[=get_type]{get_type()}} to access specific properties,
\code{\link[=load_synonyms]{load_synonyms()}}, \code{\link[=get_all_labels]{get_all_labels()}}, \code{\link[=load_redirections]{load_redirections()}}, \code{\link[=load_redirections]{load_redirections()}} to access other pieces of information contained in the nomenclature file.
}
