% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{load_nomenclature}
\alias{load_nomenclature}
\title{Load the nomenclature file given in the Orphanet nomenclature pack,
then process it to make it more R-friendly.}
\usage{
load_nomenclature()
}
\value{
The nomenclature data as a data.frame object.
}
\description{
Load the nomenclature file given in the Orphanet nomenclature pack,
then process it to make it more R-friendly.
}
\examples{
nom_data = load_nomenclature()

}
