% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{load_synonyms}
\alias{load_synonyms}
\title{Load for each orphaCode its list of synonyms from the nomenclature data}
\usage{
load_synonyms()
}
\value{
A data.frame object with three columns : ORPHAcode, preferred term, synonyms
}
\description{
Load for each orphaCode its list of synonyms from the nomenclature data
}
\examples{
df_synonyms = load_synonyms()
}
