% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{merge_graphs}
\alias{merge_graphs}
\title{Convert a list of graphs into a single merged graph}
\usage{
merge_graphs(graphs_list)
}
\arguments{
\item{graphs_list}{The graphs to merge}
}
\value{
The merged graph
}
\description{
Convert a list of graphs into a single merged graph
}
\examples{
code = 303
df_descendants = get_descendants(code)
df_ancestors = get_ancestors(code)
merged_graph = merge_graphs(list(df_descendants, df_ancestors))
}
