% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{merge_graphs}
\alias{merge_graphs}
\title{Merge graphs}
\usage{
merge_graphs(graphs_list)
}
\arguments{
\item{graphs_list}{The graphs to merge.}
}
\value{
The merged graph
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}

Convert a list of graphs into a single merged graph
}
\examples{
code = '303'
graph_descendants = get_descendants(code, output='graph')
graph_ancestors = get_ancestors(code, output='graph')

merged_graph = merge_graphs(list(graph_descendants, graph_ancestors))
}
