% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\name{orpha_df}
\alias{orpha_df}
\title{Prepare grouping}
\usage{
orpha_df(
  x,
  orpha_code_col = "orpha_code",
  df_classif = NULL,
  force_codes = NULL
)
}
\arguments{
\item{x}{A data.frame object or data.frame extension.}

\item{orpha_code_col}{The column name of \code{x} containing the ORPHAcodes.}

\item{df_classif}{The classification to consider, in a from/to format.}

\item{force_codes}{ORPHAcodes to be forced to appear in \code{groups} attribute after a \code{\link[=group_by]{group_by()}} operation.}
}
\value{
An \code{orpha_df} instantiation. Any data.frame extension will be kept.
Former attributes are erased if \code{orpha_df} is called multiple times.
}
\description{
This class allows performing hierarchical grouping operations using the Orphanet classification system.
Basically each ORPHAcode group built with \code{\link[=group_by.orpha_df]{group_by.orpha_df()}} will also contain rows with descendants ORPHAcodes.
This away a row may belong to several groups.

The \code{orpha_df} class just needs to know in which column ORPHAcodes are located to work.
}
\examples{
library(dplyr)

# Build patients data.frame
df_patients = data.frame(
patient_id = c(1,1,2,3,4,5,6),
code = c('303', '158673', '595356', '305', '79406', '79406', '595356'))

df_counts = df_patients \%>\%
 group_by(code) \%>\%
 count() \%>\%
 as.data.frame()

df_counts = df_patients \%>\%
 orpha_df(orpha_code_col = 'code') \%>\%
 group_by(code) \%>\%
 count() \%>\%
 as.data.frame()

}
