% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplication.R
\name{roll_up_descendants}
\alias{roll_up_descendants}
\title{For counting operations, all descendants of a given code should be rolled up.
The rows corresponding to a descendant are then duplicated to be attributed to the given code also}
\usage{
roll_up_descendants(.data, orphaCode, edgelist, code_col = "code")
}
\arguments{
\item{.data}{The data to consider}

\item{orphaCode}{The given Orpha code}

\item{edgelist}{Information of how codes are related to each other}
}
\value{
The duplicated descendants rows newly attributed to the given Orpha code
}
\description{
For counting operations, all descendants of a given code should be rolled up.
The rows corresponding to a descendant are then duplicated to be attributed to the given code also
}
