% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{subtype_to_disorder}
\alias{subtype_to_disorder}
\title{Find the associated disorder Orpha code for a given subtype code}
\usage{
subtype_to_disorder(subtypeCode, class_data = NULL)
}
\arguments{
\item{subtypeCode}{The Orpha code of the subtype of disorder}

\item{class_data}{The classification to consider.
If NULL, automatically search a classification for each code}
}
\value{
The Orpha code of the associated disorder
}
\description{
Find the associated disorder Orpha code for a given subtype code
}
\examples{
subtype_to_disorder(subtypeCode = 158676) # 158676 is a subtype of disorder
subtype_to_disorder(subtypeCode = 303) # 303 is a group of disorder
}
