% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{translate_class_level}
\alias{translate_class_level}
\title{Translate a classLevel column into human readable data}
\usage{
translate_class_level(V)
}
\arguments{
\item{V, }{Vector to translate}
}
\value{
The dataframe with the translated column
}
\description{
Translate a classLevel column into human readable data
}
