% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{translate_disorder_types}
\alias{translate_disorder_types}
\title{Translate a disorderType column into human readable data}
\usage{
translate_disorder_types(df, col = "disorderType")
}
\arguments{
\item{df, }{The dataframe containing the disorderType column}

\item{col, }{To change the disorderType column name if necessary}
}
\value{
The dataframe with the translated column
}
\description{
Translate a disorderType column into human readable data
}
