% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{translate_flag_values}
\alias{translate_flag_values}
\title{Translate a flagValue column into human readable data}
\usage{
translate_flag_values(V)
}
\arguments{
\item{V, }{Vector to translate}
}
\value{
The dataframe with the translated column
}
\description{
Translate a flagValue column into human readable data
}
